Make User.fixUp more efficient by adding a Datastore.isJIDInChannel method that does not do attachUsersToChannel.
Fix "can't operate on multiple entity groups in a single transaction" when fixing up users by ending the closing/opening the persistence manager in User.

I _think_ this is necessary because of this bit from http://code.google.com/appengine/docs/java/datastore/queriesandindexes.html: "Queries are only supported inside transactions if they include an ancestor filter. The query's ancestor must be in the same entity group as the other operations in the transaction" If we ended up doing a put() in User.fixUp, that would automatically start a transaction, and subsequent User queries (in attachUsersToChannel/getUsersByChannel) would not have this relationship.