Hopefully fix User inconsistencies.
Member was marked as both @PersistenceCapable and as implementing Serializable. While the intention was that the Channel and User fields should not be persisted (as tagged with @NotPersistent), since the Set of Members in Channel is tagged with @Persistent(serialized = "true"), it was being persisted with Java serialization, not AppEngine/JDO. This meant that they were being persisted, these stale User objects were also occasionally being written to the User table too.

To actually mark User and Channel as not persistent for the purposes of Java serialization, use the transient keyword (and remove the JDO annotations, since they're ignored).