Add memcache caching of User and Channel objects, so that we can hopefully avoid 2 datastore reads for every incoming XMPP message. This is done by adding a CachingDatastore wrapper for Datastore, and a MemcacheCachingDatasore implementation of the wrapper (there's also an InMemoryCachingDatastore that's used for tests).
The cache keys things off of JID (for User) and name (for Channel). It adds
things to the cache on gets and puts, and removes on deletes. I initially tried
a more conservative approach where we remove on put (to refresh on the next
get), but since we do a put for the Channel on every request (to persist the
last seen field for members and sequence ID), we were not ending up with 
caching for any Channel data.