Start of a "live" web UI that uses the Channel API (I'm avoiding having too many channel references in the class names so that there won't be confusion between our channels and App Engine's).
When a user opens the live UI, and every 5 minutes after that, we record a
live "last seen" time, and then in Channel.sendMessage we send messages via
the Channel API for members that have been seen in the past 10 minutes. There's
also an unload event handler that clears the last seen time (we don't rely on
it and require the pings/heartbeats in case a user's tab crashes or otherwise
doesn't fire the unload event).

There's still lots to be done (message formatting, timestamps, etc.), so the
live UI isn't linked from anywhere.